Rownanie kwadratowe (zmienne przecinkowe)

+++ rownanie.c
@@ -1,11 +1,11 @@
 # include <stdio.h>
 # include <math.h>
-int a,b,c;
-int delta,x1,x2;
-int main()
+float a,b,c;
+float  delta,x1,x2;
+float main()
 {
 printf("Podja wartosc a: ");
-scanf("%i",&a);
+scanf("%f",&a);
 if(a==0)
 {
 printf("Podana wartość dla a nie moze wynosic 0 !!!\n");
@@ -13,20 +13,20 @@ printf("Podana wartość dla a nie moze wynosic 0 !!!\n");
 else
 {
 printf("Podaj wartosc b: ");
-scanf("%i",&b);
+scanf("%f",&b);
 printf("Podaj wartosc c: ");
-scanf("%i",&c);
+scanf("%f",&c);
 delta=0;
 delta=(b*b)-(4*a*c);
-printf("Delta wynosi: %i\n",delta);
+printf("Delta wynosi: %f\n",delta);
   if (delta>0)
 	{
 	delta=sqrt(delta);
 	x1=((-b-delta)/2*a);
 	x2=((-b+delta)/2*a);
 	printf("Rownanie kwadratowe ma dwa pierwiastki: \n");
-	printf("X1: %i\n",x1);
-	printf("X2: %i\n",x2);
+	printf("X1: %f\n",x1);
+	printf("X2: %f\n",x2);
 	}
 	else
 	{
@@ -34,7 +34,7 @@ printf("Delta wynosi: %i\n",delta);
 		{
 		x1=-b/2*a;
 		printf("Rownanie kwadratowe ma jeden pierwiastek: \n");
-		printf("X1: %i\n",x1);
+		printf("X1: %f\n",x1);
 		}
 		else
 		{
